+++ NEWS.md
@@ -3,6 +3,7 @@
 ## Bug fixes
 
 * Fix missing "testthat" dependency
+* Remove dependency on "mHealthR" package
 
 # MIMSunit 0.6.2